Change config to work for *all* examples

+++ examples/NMEAtest/NMEAtest.ino
@@ -127,7 +127,7 @@ void setup()
   uint8_t validGGA_len = 0;
 
   // Insert a ' ' at each position of the test sentence
-  uint8_t insert_at=1;
+  uint16_t insert_at=1;
   do {
     const char *ptr = validGGA;
     uint8_t j = 0;
@@ -160,7 +160,7 @@ void setup()
   passed++;
 
   // Drop one character from each position in example sentence
-  for (uint8_t i=0; i < validGGA_len-3; i++) {
+  for (uint16_t i=0; i < validGGA_len-3; i++) {
     const char *ptr = validGGA;
     uint8_t j = 0;
     char dropped = 0;
@@ -182,7 +182,7 @@ void setup()
   passed++;
 
   // Mangle one character from each position in example sentence
-  for (uint8_t i=0; i < validGGA_len-3; i++) {
+  for (uint16_t i=0; i < validGGA_len-3; i++) {
     const char *ptr = validGGA;
     uint8_t j = 0;
     char replaced = 0;
